feat: update theme and add bar status style
