+++ packages/frontend/mobile/src/components/StatusBarMobile.tsx
@@ -1,0 +1,13 @@
+import { StatusBar } from 'react-native';
+
+import { useTheme } from '../contexts/ThemeContext';
+
+export default function StatusBarMobile() {
+  const { colorTheme } = useTheme();
+
+  return (
+    <StatusBar
+      barStyle={colorTheme === 'dark' ? 'light-content' : 'dark-content'}
+    />
+  );
+}
